requirements-test.txt
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
testcontainers>=3.7.0
factory-boy>=3.3.0
freezegun>=1.2.0